+++ tupan/particles/body.py
@@ -10,8 +10,9 @@ import sys
 import copy
 import numpy as np
 from ..lib import gravity
-from ..lib.utils.timing import decallmethods, timings
 from ..lib.utils.memoize import cache, cache_arg
+from ..lib.utils.timing import decallmethods, timings
+from ..lib.utils.dtype import *
 
 
 __all__ = ["Bodies"]
@@ -71,18 +72,18 @@ class NbodyMethods(NbodyUtils):
     This class holds common methods for particles in n-body systems.
     """
     attrs = [# name, dtype, doc
-             ("id", np.uint64, "index"),
-             ("mass", np.float64, "mass"),
-             ("eps2", np.float64, "squared softening"),
-             ("x", np.float64, "x-position"),
-             ("y", np.float64, "y-position"),
-             ("z", np.float64, "z-position"),
-             ("vx", np.float64, "x-velocity"),
-             ("vy", np.float64, "y-velocity"),
-             ("vz", np.float64, "z-velocity"),
-             ("time", np.float64, "current time"),
-             ("tstep", np.float64, "time step"),
-             ("nstep", np.uint64, "step number"),
+             ("id", UINT, "index"),
+             ("mass", REAL, "mass"),
+             ("eps2", REAL, "squared softening"),
+             ("x", REAL, "x-position"),
+             ("y", REAL, "y-position"),
+             ("z", REAL, "z-position"),
+             ("vx", REAL, "x-velocity"),
+             ("vy", REAL, "y-velocity"),
+             ("vz", REAL, "z-velocity"),
+             ("time", REAL, "current time"),
+             ("tstep", REAL, "time step"),
+             ("nstep", UINT, "step number"),
             ]
     names = [_[0] for _ in attrs]
     dtype = [(_[0], _[1]) for _ in attrs]
@@ -358,19 +359,19 @@ class PNbodyMethods(NbodyMethods):
     This class holds common methods for particles in n-body systems with post-Newtonian corrections.
     """
     special_attrs = [# name, dtype, doc
-                     ("pn_dvx", np.float64, "post-Newtonian correction for the x-velocity"),
-                     ("pn_dvy", np.float64, "post-Newtonian correction for the y-velocity"),
-                     ("pn_dvz", np.float64, "post-Newtonian correction for the z-velocity"),
-                     ("pn_ke", np.float64, "post-Newtonian correction for the kinetic energy"),
-                     ("pn_rcomx", np.float64, "post-Newtonian correction for the center-of-mass x-position"),
-                     ("pn_rcomy", np.float64, "post-Newtonian correction for the center-of-mass y-position"),
-                     ("pn_rcomz", np.float64, "post-Newtonian correction for the center-of-mass z-position"),
-                     ("pn_lmx", np.float64, "post-Newtonian correction for the x-linear momentum"),
-                     ("pn_lmy", np.float64, "post-Newtonian correction for the y-linear momentum"),
-                     ("pn_lmz", np.float64, "post-Newtonian correction for the z-linear momentum"),
-                     ("pn_amx", np.float64, "post-Newtonian correction for the x-angular momentum"),
-                     ("pn_amy", np.float64, "post-Newtonian correction for the y-angular momentum"),
-                     ("pn_amz", np.float64, "post-Newtonian correction for the z-angular momentum"),
+                     ("pn_dvx", REAL, "post-Newtonian correction for the x-velocity"),
+                     ("pn_dvy", REAL, "post-Newtonian correction for the y-velocity"),
+                     ("pn_dvz", REAL, "post-Newtonian correction for the z-velocity"),
+                     ("pn_ke", REAL, "post-Newtonian correction for the kinetic energy"),
+                     ("pn_rcomx", REAL, "post-Newtonian correction for the center-of-mass x-position"),
+                     ("pn_rcomy", REAL, "post-Newtonian correction for the center-of-mass y-position"),
+                     ("pn_rcomz", REAL, "post-Newtonian correction for the center-of-mass z-position"),
+                     ("pn_lmx", REAL, "post-Newtonian correction for the x-linear momentum"),
+                     ("pn_lmy", REAL, "post-Newtonian correction for the y-linear momentum"),
+                     ("pn_lmz", REAL, "post-Newtonian correction for the z-linear momentum"),
+                     ("pn_amx", REAL, "post-Newtonian correction for the x-angular momentum"),
+                     ("pn_amy", REAL, "post-Newtonian correction for the y-angular momentum"),
+                     ("pn_amz", REAL, "post-Newtonian correction for the z-angular momentum"),
                     ]
     special_names = [_[0] for _ in special_attrs]
     special_dtype = [(_[0], _[1]) for _ in special_attrs]
@@ -574,7 +575,7 @@ class Bodies(AbstractNbodyMethods):
     """
     def __init__(self, n=0, items=None):
         if items is None:
-            self.__dict__['id'] = np.arange(n, dtype=np.uint64)
+            self.__dict__['id'] = np.arange(n, dtype=UINT)
             for name, dtype in self.dtype+self.special_dtype:
                 self.__dict__[name] = np.zeros(n, dtype=dtype)
         else:
