+++ tupan/particles/star.py
@@ -9,6 +9,7 @@ from __future__ import print_function
 import numpy as np
 from .body import Bodies
 from ..lib.utils.timing import decallmethods, timings
+from ..lib.utils.dtype import *
 
 
 __all__ = ["Stars"]
@@ -20,12 +21,12 @@ class Stars(Bodies):
 
     """
     dtype = Bodies.dtype + [
-                            ("sx", np.float64),
-                            ("sy", np.float64),
-                            ("sz", np.float64),
-                            ("radius", np.float64),
-                            ("age", np.float64),
-                            ("metallicity", np.float64),
+                            ("sx", REAL),
+                            ("sy", REAL),
+                            ("sz", REAL),
+                            ("radius", REAL),
+                            ("age", REAL),
+                            ("metallicity", REAL),
                            ]
